+++ FirestoreExtractor/firestore_extractor.py
@@ -3,6 +3,17 @@ from firebase_admin import *
 from firebase_admin import firestore
 import json
 
+from firebase_admin import credentials
+from firebase_admin import db
+
+
+def  extractFromFirebase(certificate_json_file_path,firebase_url):
+    cred = credentials.Certificate(certificate_json_file_path)
+    firebase_admin.initialize_app(cred,{'databaseURL':firebase_url})
+    ref=db.reference()
+    mydict=ref.get()
+    return mydict
+
 
 def extractFromFirstProjectOfFirebase(certificate_json_file_path):
     cred = credentials.Certificate(certificate_json_file_path)
